requirements.txt
django
djangorestframework
djangorestframework-jwt
django-phonenumber-field
tox